pyproject.toml
[tool.poetry]
name = "llm-for-sql"
version = "0.1.0"
description = ""
authors = ["thanktua-spp <egbethankgodjnr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.13"
langchain = "^0.1.11"
openai = "^1.13.3"
fastapi = "^0.110.0"
tiktoken = "^0.6.0"
uvicorn = "^0.28.0"
pydantic = "^2.6.3"
python-dotenv = "^1.0.1"
langchainhub = "^0.1.15"
langchain-community = "^0.0.27"
langchain-openai = "^0.0.8"
pinecone-client = "^3.1.0"
pypdf = "^4.1.0"
langchain-pinecone = "^0.0.3"
langserve = {extras = ["all"], version = "^0.0.51"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
mypy = "^1.9.0"
ipykernel = "^6.29.3"
jupyter = "^1.0.0"
faiss-cpu = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
